refactor: 뷰 분리 및 열거형으로 각 섹션 처리
