+++ SecondSeminar/SecondSeminar/Sources/Presentation/Enum/AppDetailSection.swift
@@ -1,0 +1,17 @@
+//
+//  AppDetailSection.swift
+//  SecondSeminar
+//
+//  Created by RAFA on 10/14/24.
+//
+
+import Foundation
+
+enum AppDetailSection: CaseIterable {
+    case header
+    case metaInfo
+    case whatsNew
+    case preview
+    case developerInfo
+    case feedback
+}
